backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
PyPDF2==3.0.1
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
groq==0.11.0
